requirements.txt
dash==2.2.0
dash_bootstrap_components==1.0.2
plotly_express==0.4.0
jupyter-dash==0.4.1
pandas==1.3.2
py4j==0.10.9
pyarrow==5.0.0
pyspark==3.1.2
koalas==1.8.1
openpyxl==3.0.7
urllib3==1.26.6
websocket-client==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.1
SQLAlchemy==1.4.23
databases==0.5.0
asyncpg==0.24.0
psycopg2-binary==2.9.1
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
dash_extensions==0.0.60
kaleido
visdcc==0.0.40
jupyterlab==3.2.9
dash-labs==1.0.2
dash-leaflet==0.1.23
jsbeautifier
apache-sedona==1.1.1

